wip: add util and hooks
